complete 部门管理页面中岗位显示 完成job service层代码

+++ src/main/java/com/hrms/dao/JobMapper.java
@@ -18,7 +18,7 @@ public interface JobMapper {
 
     List<Job> selectByExample(JobExample example);
 
-    Job selectByPrimaryKey(Short jobid);
+    Job selectByPrimaryKey(Integer jobid);
 
     int updateByExampleSelective(@Param("record") Job record, @Param("example") JobExample example);
